refactoring: moved fetchManifest to API folder
